+++ Assets/Scripts/ScriptableObjects/Stats.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0304f522074b34343928303d59d211ea
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
