+++ Assets/Scripts/ScriptableObjects/Stats/BaseStatSO.cs.meta
